+++ examples/integration_test_example/lib/widgetbook.config.dart
@@ -2,6 +2,4 @@ import 'package:widgetbook/widgetbook.dart';
 
 import 'components.g.dart';
 
-final config = Config(
-  components: components,
-);
+final config = Config(components: components);